+++ test/dummy/app/helpers/application_helper.rb
@@ -57,4 +57,25 @@ module ApplicationHelper
       recording: current_root_recording
     )
   end
+
+  def dummy_show_press_kits?
+    RecordingStudioStripe::Limits.known?(:press_kits) && current_root_recording&.recordable.is_a?(Workspace)
+  end
+
+  def dummy_press_kit_recordings
+    return RecordingStudio::Recording.none unless current_root_recording
+
+    RecordingStudio::Recording.for_root(current_root_recording.id).of_type("PressKit").where(trashed_at: nil).order(:created_at)
+  end
+
+  def dummy_press_kits_subtitle
+    return "Pick a plan to add press kits." unless current_root_recording&.recordable.respond_to?(:billing)
+
+    handle = current_root_recording.recordable.billing.limit(:press_kits)
+    if handle.included <= 0
+      "Pick a plan to add press kits."
+    else
+      "#{handle.used} of #{handle.included} on this plan."
+    end
+  end
 end